backend/requirements.txt
mysql-connector-python
bcrypt
flask